Makes the enter key search for the first available option. Also cleans up the autocomplete
